+++ src/main/java/ca/uwaterloo/drinkmasterapi/common/CurrencyEnum.java
@@ -1,0 +1,7 @@
+package ca.uwaterloo.drinkmasterapi.common;
+
+public enum CurrencyEnum {
+    CAD,
+    USD,
+    CNY
+}
