+++ src/main/java/ca/uwaterloo/drinkmasterapi/dao/DrinkIngredient.java
@@ -1,4 +1,4 @@
-package ca.uwaterloo.drinkmasterapi.feature.drink.model;
+package ca.uwaterloo.drinkmasterapi.dao;
 
 import lombok.Getter;
 import lombok.Setter;
@@ -8,22 +8,25 @@ import javax.persistence.*;
 import java.time.OffsetDateTime;
 
 @Entity
-@Table(name = "ingredient_allergen")
+@Table(name = "drink_ingredient")
 @Getter
 @Setter
 @ToString
-public class IngredientAllergen {
+public class DrinkIngredient {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
     @ManyToOne
+    @JoinColumn(name = "drink_id")
+    private Drink drink;
+
+    @ManyToOne
     @JoinColumn(name = "ingredient_id")
     private Ingredient ingredient;
 
-    @ManyToOne
-    @JoinColumn(name = "allergen_id")
-    private Allergen allergen;
+    @Column(name = "quantity", nullable = false, columnDefinition = "DOUBLE DEFAULT 0.0")
+    private Double quantity;
 
     @Column(name = "created_at", nullable = false, updatable = false, columnDefinition = "TIMESTAMP DEFAULT CURRENT_TIMESTAMP")
     private OffsetDateTime createdAt;
