+++ src/main/java/ca/uwaterloo/drinkmasterapi/dao/Machine.java
@@ -1,4 +1,4 @@
-package ca.uwaterloo.drinkmasterapi.feature.mqtt.model;
+package ca.uwaterloo.drinkmasterapi.dao;
 
 import lombok.Data;
 import lombok.Getter;
