+++ src/main/java/ca/uwaterloo/drinkmasterapi/dao/UserAllergen.java
@@ -1,6 +1,5 @@
-package ca.uwaterloo.drinkmasterapi.feature.user.model;
+package ca.uwaterloo.drinkmasterapi.dao;
 
-import ca.uwaterloo.drinkmasterapi.feature.drink.model.Allergen;
 import lombok.Getter;
 import lombok.Setter;
 import lombok.ToString;
