--- src/main/java/ca/uwaterloo/drinkmasterapi/feature/order/model/CurrencyEnum.java
@@ -1,7 +1,0 @@
-package ca.uwaterloo.drinkmasterapi.feature.order.model;
-
-public enum CurrencyEnum {
-    CAD,
-    USD,
-    CNY
-}
