+++ src/main/java/ca/uwaterloo/drinkmasterapi/feature/order/repository/OrderRepository.java
@@ -1,6 +1,6 @@
 package ca.uwaterloo.drinkmasterapi.feature.order.repository;
 
-import ca.uwaterloo.drinkmasterapi.feature.order.model.Order;
+import ca.uwaterloo.drinkmasterapi.dao.Order;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
