+++ src/main/java/ca/uwaterloo/drinkmasterapi/feature/user/repository/UserRepository.java
@@ -1,6 +1,6 @@
 package ca.uwaterloo.drinkmasterapi.feature.user.repository;
 
-import ca.uwaterloo.drinkmasterapi.feature.user.model.User;
+import ca.uwaterloo.drinkmasterapi.dao.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
