+++ src/main/java/ca/uwaterloo/drinkmasterapi/feature/user/service/UserLoginServiceImpl.java
@@ -4,7 +4,7 @@ import ca.uwaterloo.drinkmasterapi.common.InvalidCredentialsException;
 import ca.uwaterloo.drinkmasterapi.feature.user.model.LoginRequestDTO;
 import ca.uwaterloo.drinkmasterapi.feature.user.model.UserResponseDTO;
 import ca.uwaterloo.drinkmasterapi.feature.user.model.SignupRequestDTO;
-import ca.uwaterloo.drinkmasterapi.feature.user.model.User;
+import ca.uwaterloo.drinkmasterapi.dao.User;
 import ca.uwaterloo.drinkmasterapi.feature.user.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.crypto.password.PasswordEncoder;
