+++ src/test/java/ca/uwaterloo/drinkmasterapi/feature/user/UserLoginServiceImplTest.java
@@ -3,7 +3,7 @@ package ca.uwaterloo.drinkmasterapi.feature.user;
 import ca.uwaterloo.drinkmasterapi.common.InvalidCredentialsException;
 import ca.uwaterloo.drinkmasterapi.feature.user.model.LoginRequestDTO;
 import ca.uwaterloo.drinkmasterapi.feature.user.model.SignupRequestDTO;
-import ca.uwaterloo.drinkmasterapi.feature.user.model.User;
+import ca.uwaterloo.drinkmasterapi.dao.User;
 import ca.uwaterloo.drinkmasterapi.feature.user.model.UserResponseDTO;
 import ca.uwaterloo.drinkmasterapi.feature.user.repository.UserRepository;
 import ca.uwaterloo.drinkmasterapi.feature.user.service.UserLoginServiceImpl;
